requirements.txt
snowflake-connector-python==3.5.0
python-dotenv==1.0.0
pandas==2.1.4
streamlit==1.29.0
numpy==1.26.2
plotly==5.18.0 